Some warning fixes.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@32081 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/actionreplay.c
@@ -214,7 +214,7 @@ static int actionreplay_dump(void)
     mon_out("EXROM line: %s, GAME line: %s, Mode: %s\n",
             (regvalue & 2) ? "high" : "low",
             (regvalue & 1) ? "low" : "high",
-            cart_config_string(regvalue & 3));
+            cart_config_string((BYTE)(regvalue & 3)));
     mon_out("ROM bank: %d, cart state: %s, reset freeze: %s\n",
             (regvalue & 0x18) >> 3,
             (regvalue & 4) ? "disabled" : "enabled",
